requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
torch>=2.0.0
pytorch-lightning>=2.0.0
ta>=0.10.2
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0